requirements.txt
asgiref==3.11.1
Django==6.0.3
django-modeltranslation
gunicorn==23.0.0
pillow==12.1.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
